+++ app/frontend/src/javascript/components/machines/reserve-button.tsx
@@ -46,7 +46,7 @@ const ReserveButton: React.FC<ReserveButtonProps> = ({ currentUser, machineId, o
   // check the trainings after we retrieved the machine data
   useEffect(() => checkTraining(), [machine]);
   useEffect(() => {
-    SettingAPI.get(SettingName.PackOnlyForSubscription)
+    SettingAPI.get('pack_only_for_subscription')
       .then(data => setIsPackOnlyForSubscription(data.value === 'true'))
       .catch(error => onError(error));
   }, []);