+++ app/frontend/src/javascript/components/payment/stripe/stripe-elements.tsx
@@ -14,7 +14,7 @@ export const StripeElements: React.FC = memo(({ children }) => {
    * When this component is mounted, we initialize the <Elements> tag with the Stripe's public key
    */
   useEffect(() => {
-    SettingAPI.get(SettingName.StripePublicKey).then(key => {
+    SettingAPI.get('stripe_public_key').then(key => {
       if (key?.value) {
         const promise = loadStripe(key.value);
         setStripe(promise);