+++ app/frontend/src/javascript/components/payment/stripe/stripe-keys-form.tsx
@@ -42,9 +42,9 @@ const StripeKeysForm: React.FC<StripeKeysFormProps> = ({ onValidKeys, onInvalidK
   useEffect(() => {
     mounted.current = true;
 
-    SettingAPI.query([SettingName.StripePublicKey, SettingName.StripeSecretKey]).then(stripeKeys => {
-      setPublicKey(stripeKeys.get(SettingName.StripePublicKey));
-      setSecretKey(stripeKeys.get(SettingName.StripeSecretKey));
+    SettingAPI.query(['stripe_public_key', 'stripe_secret_key']).then(stripeKeys => {
+      setPublicKey(stripeKeys.get('stripe_public_key'));
+      setSecretKey(stripeKeys.get('stripe_secret_key'));
     }).catch(error => console.error(error));
 
     // when the component unmounts, mark it as unmounted